better makefile

+++ src/buffer.c
@@ -1,7 +1,7 @@
 #include "buffer.h"
 
 void ab_append(append_buf* ab, const char* s, int len) {
-  char *new = realloc(ab->b, ab->len + len);
+  char *new = realloc(ab->b, (unsigned long) ab->len + (unsigned long) len);
   if (new == NULL) return;
   memcpy(&new[ab->len], s, len);
   ab->b = new;
